+++ mocks/mocks-server.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   options: {
-    port: 3200
-  }
+    port: 3200,
+  },
 };