+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 ### Changed
+- chore(lint): Lint mocks folder
 ### Fixed
 - fix: Remove mocks-server-plugin tag from package.json (#93)
 ### Removed