+++ CHANGELOG.md
@@ -10,14 +10,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+### Changed
+### Fixed
+### Removed
+
+## [1.5.0] - 2020-12-25
+
+### Added
 - feat(#78): Add alerts router
 - chore(lint): Add eslint plugin to avoid only in tests
+
 ### Changed
 - chore(deps): Update dependencies
 - test(e2e): Move utils to a support folder
 - feat: Add/remove routers on start/stop plugin methods, not in init method.
-### Fixed
-### Removed
 
 ## [1.4.7] - 2020-12-21